Lots of changes
Primarily, writing up based on experience of trying to implement this twice:
- once just stabilizing the existing `Deref` approach
- once by adding the `Receiver` trait and a blanket impl for `Deref`

plus just lots of thinking and writing about it.

The only major substantive change here is that I propose we do NOT
stabilize the use of raw pointers as self types. For the reasons
described, I don't see any path to ever stabilizing this, and we
should remove it.